Move back methods to StringSupport
