+++ src/java.base/share/classes/jdk/internal/foreign/SegmentBulkOperations.java
@@ -378,198 +378,6 @@ public final class SegmentBulkOperations {
         return ~remaining;
     }
 
-    //private static final HexFormat HF = HexFormat.ofDelimiter(" ");
-
-    /**
-     * {@return the shortest distance beginning at the provided {@code fromOffset}
-     *          to the encountering of a zero byte in the provided {@code segment}
-     *          checking bytes before the {@code toOffset}}
-     * <p>
-     * The method is using a heuristic method to determine if a long word contains a
-     * zero byte. The method might have false positives but never false negatives.
-     * <p>
-     * This method is inspired by the `glibc/string/strlen.c` implementation
-     *
-     * @param segment    to examine
-     * @param fromOffset from where examination shall begin (inclusive)
-     * @param toOffset   to where examination shall end (exclusive)
-     * @throws IllegalArgumentException if the examined region contains no zero bytes
-     *                                  within a length that can be accepted by a String
-     */
-    @ForceInline
-    public static int strlenByte(final AbstractMemorySegmentImpl segment,
-                                 final long fromOffset,
-                                 final long toOffset) {
-        final long length = toOffset - fromOffset;
-        segment.checkBounds(fromOffset, length);
-        if (length == 0) {
-            // The state has to be checked explicitly for zero-length segments
-            segment.scope.checkValidState();
-            throw stringTooLarge(segment, fromOffset, toOffset);
-        }
-        final long longBytes = length & LONG_MASK;
-        final long longLimit = fromOffset + longBytes;
-        long offset = fromOffset;
-        for (; offset < longLimit; offset += Long.BYTES) {
-            long val = SCOPED_MEMORY_ACCESS.getLongUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset, !Architecture.isLittleEndian());
-            if (mightContainZeroByte(val)) {
-                for (int j = 0; j < Long.BYTES; j++) {
-                    if (SCOPED_MEMORY_ACCESS.getByte(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset + j) == 0) {
-                        return requireWithinArraySize(offset + j - fromOffset, segment, fromOffset, toOffset);
-                    }
-                }
-            }
-        }
-        // Handle the tail
-        for (; offset < toOffset; offset++) {
-            byte val = SCOPED_MEMORY_ACCESS.getByte(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset);
-            if (val == 0) {
-                return requireWithinArraySize(offset - fromOffset, segment, fromOffset, toOffset);
-            }
-        }
-        throw nullNotFound(segment, fromOffset, toOffset);
-    }
-
-    @ForceInline
-    public static int strlenShort(final AbstractMemorySegmentImpl segment,
-                                  final long fromOffset,
-                                  final long toOffset) {
-        final long length = toOffset - fromOffset;
-        segment.checkBounds(fromOffset, length);
-        if (length == 0) {
-            segment.scope.checkValidState();
-            throw stringTooLarge(segment, fromOffset, toOffset);
-        }
-        final long longBytes = length & LONG_MASK;
-        final long longLimit = fromOffset + longBytes;
-        long offset = fromOffset;
-        for (; offset < longLimit; offset += Long.BYTES) {
-            long val = SCOPED_MEMORY_ACCESS.getLongUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset, !Architecture.isLittleEndian());
-            if (mightContainZeroShort(val)) {
-                for (int j = 0; j < Long.BYTES; j += Short.BYTES) {
-                    if (SCOPED_MEMORY_ACCESS.getShortUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset + j, !Architecture.isLittleEndian()) == 0) {
-                        return requireWithinArraySize(offset + j - fromOffset, segment, fromOffset, toOffset);
-                    }
-                }
-            }
-        }
-        // Handle the tail
-        // Prevent over scanning as we step by 2
-        final long endScan = toOffset & ~1; // The last bit is zero
-        for (; offset < endScan; offset += Short.BYTES) {
-            short val = SCOPED_MEMORY_ACCESS.getShortUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset, !Architecture.isLittleEndian());
-            if (val == 0) {
-                return requireWithinArraySize(offset - fromOffset, segment, fromOffset, toOffset);
-            }
-        }
-        throw nullNotFound(segment, fromOffset, toOffset);
-    }
-
-    @ForceInline
-    public static int strlenInt(final AbstractMemorySegmentImpl segment,
-                                final long fromOffset,
-                                final long toOffset) {
-        final long length = toOffset - fromOffset;
-        segment.checkBounds(fromOffset, length);
-        if (length == 0) {
-            segment.scope.checkValidState();
-            throw stringTooLarge(segment, fromOffset, toOffset);
-        }
-        final long longBytes = length & LONG_MASK;
-        final long longLimit = fromOffset + longBytes;
-        long offset = fromOffset;
-        for (; offset < longLimit; offset += Long.BYTES) {
-            long val = SCOPED_MEMORY_ACCESS.getLongUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset, !Architecture.isLittleEndian());
-            if (mightContainZeroShort(val)) {
-                for (int j = 0; j < Long.BYTES; j += Integer.BYTES) {
-                    if (SCOPED_MEMORY_ACCESS.getIntUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset + j, !Architecture.isLittleEndian()) == 0) {
-                        return requireWithinArraySize(offset + j - fromOffset, segment, fromOffset, toOffset);
-                    }
-                }
-            }
-        }
-        // Handle the tail
-        // Prevent over scanning as we step by 4
-        final long endScan = toOffset & ~3; // The last two bit are zero
-        for (; offset < endScan; offset += Integer.BYTES) {
-            int val = SCOPED_MEMORY_ACCESS.getIntUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset, !Architecture.isLittleEndian());
-            if (val == 0) {
-                return requireWithinArraySize(offset - fromOffset, segment, fromOffset, toOffset);
-            }
-        }
-        throw nullNotFound(segment, fromOffset, toOffset);
-    }
-
-
-/*    // The gain of using `long` wide operations for `int` is lower than for the two other
-    // `byte` and `short` variants so, there is only a simpler method for `int`s
-    @ForceInline
-    public static int strlenInt(AbstractMemorySegmentImpl segment, long fromOffset, long toOffset) {
-        long length = Math.min(toOffset - fromOffset, ArraysSupport.SOFT_MAX_ARRAY_LENGTH) & ~3;
-        for (int offset = 0; offset < length; offset += Integer.BYTES) {
-            // We are guaranteed to be aligned here so, we can use unaligned access.
-            int val = SCOPED_MEMORY_ACCESS.getIntUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset, !Architecture.isLittleEndian());
-            if (val == 0) {
-                return offset;
-            }
-        }
-        throw nullNotFound(segment, fromOffset, toOffset);
-    }*/
-
-    /*
-    Bits 63 and N * 8 (N = 1..7) of this number are zero.  Call these bits
-    the "holes".  Note that there is a hole just to the left of
-    each byte, with an extra at the end:
-
-    bits:  01111110 11111110 11111110 11111110 11111110 11111110 11111110 11111111
-    bytes: AAAAAAAA BBBBBBBB CCCCCCCC DDDDDDDD EEEEEEEE FFFFFFFF GGGGGGGG HHHHHHHH
-
-    The 1-bits make sure that carries propagate to the next 0-bit.
-    The 0-bits provide holes for carries to fall into.
-    */
-    private static final long HIMAGIC_FOR_BYTES = 0x8080_8080_8080_8080L;
-    private static final long LOMAGIC_FOR_BYTES = 0x0101_0101_0101_0101L;
-
-    private static boolean mightContainZeroByte(long l) {
-        return ((l - LOMAGIC_FOR_BYTES) & (~l) & HIMAGIC_FOR_BYTES) != 0;
-    }
-
-    private static final long HIMAGIC_FOR_SHORTS = 0x8000_8000_8000_8000L;
-    private static final long LOMAGIC_FOR_SHORTS = 0x0001_0001_0001_0001L;
-
-    static boolean mightContainZeroShort(long l) {
-        return ((l - LOMAGIC_FOR_SHORTS) & (~l) & HIMAGIC_FOR_SHORTS) != 0;
-    }
-
-
-    private static int requireWithinArraySize(long size,
-                                              AbstractMemorySegmentImpl segment,
-                                              long fromOffset,
-                                              long toOffset) {
-        if (size > ArraysSupport.SOFT_MAX_ARRAY_LENGTH) {
-            throw stringTooLarge(segment, fromOffset, toOffset);
-        }
-        return (int) size;
-    }
-
-    private static IllegalArgumentException stringTooLarge(AbstractMemorySegmentImpl segment,
-                                                           long fromOffset,
-                                                           long toOffset) {
-        return new IllegalArgumentException("String too large: " + exceptionInfo(segment, fromOffset, toOffset));
-    }
-
-    private static IndexOutOfBoundsException nullNotFound(AbstractMemorySegmentImpl segment,
-                                                          long fromOffset,
-                                                          long toOffset) {
-        return new IndexOutOfBoundsException("No null terminator found: " + exceptionInfo(segment, fromOffset, toOffset));
-    }
-
-    private static String exceptionInfo(AbstractMemorySegmentImpl segment,
-                                        long fromOffset,
-                                        long toOffset) {
-        return segment + " using region [" + fromOffset + ", " + toOffset + ")";
-    }
-
     static final String PROPERTY_PATH = "java.lang.foreign.native.threshold.power.";
 
     // The returned value is in the interval [0, 2^30]
