+++ src/java.base/share/classes/jdk/internal/foreign/StringSupport.java
@@ -27,6 +27,9 @@ package jdk.internal.foreign;
 
 import jdk.internal.access.JavaLangAccess;
 import jdk.internal.access.SharedSecrets;
+import jdk.internal.misc.ScopedMemoryAccess;
+import jdk.internal.util.Architecture;
+import jdk.internal.util.ArraysSupport;
 import jdk.internal.vm.annotation.ForceInline;
 
 import java.lang.foreign.MemorySegment;
@@ -39,7 +42,9 @@ import static java.lang.foreign.ValueLayout.*;
  */
 public final class StringSupport {
 
-    static final JavaLangAccess JAVA_LANG_ACCESS = SharedSecrets.getJavaLangAccess();
+    private static final JavaLangAccess JAVA_LANG_ACCESS = SharedSecrets.getJavaLangAccess();
+    private static final ScopedMemoryAccess SCOPED_MEMORY_ACCESS = ScopedMemoryAccess.getScopedMemoryAccess();
+    private static final long LONG_MASK = ~7L; // The last three bits are zero
 
     private StringSupport() {}
 
@@ -63,7 +68,7 @@ public final class StringSupport {
 
     @ForceInline
     private static String readByte(AbstractMemorySegmentImpl segment, long offset, Charset charset) {
-        final int len = SegmentBulkOperations.strlenByte(segment, offset, segment.byteSize());
+        final int len = strlenByte(segment, offset, segment.byteSize());
         final byte[] bytes = new byte[len];
         MemorySegment.copy(segment, JAVA_BYTE, offset, bytes, 0, len);
         return new String(bytes, charset);
@@ -77,7 +82,7 @@ public final class StringSupport {
 
     @ForceInline
     private static String readShort(AbstractMemorySegmentImpl segment, long offset, Charset charset) {
-        int len = SegmentBulkOperations.strlenShort(segment, offset, segment.byteSize());
+        int len = strlenShort(segment, offset, segment.byteSize());
         byte[] bytes = new byte[len];
         MemorySegment.copy(segment, JAVA_BYTE, offset, bytes, 0, len);
         return new String(bytes, charset);
@@ -91,7 +96,7 @@ public final class StringSupport {
 
     @ForceInline
     private static String readInt(AbstractMemorySegmentImpl segment, long offset, Charset charset) {
-        int len = SegmentBulkOperations.strlenInt(segment, offset, segment.byteSize());
+        int len = strlenInt(segment, offset, segment.byteSize());
         byte[] bytes = new byte[len];
         MemorySegment.copy(segment, JAVA_BYTE, offset, bytes, 0, len);
         return new String(bytes, charset);
@@ -101,6 +106,180 @@ public final class StringSupport {
     private static void writeInt(AbstractMemorySegmentImpl segment, long offset, Charset charset, String string) {
         int bytes = copyBytes(string, segment, charset, offset);
         segment.set(JAVA_INT_UNALIGNED, offset + bytes, 0);
+    }
+
+    /**
+     * {@return the shortest distance beginning at the provided {@code fromOffset}
+     *          to the encountering of a zero byte in the provided {@code segment}
+     *          checking bytes before the {@code toOffset}}
+     * <p>
+     * The method is using a heuristic method to determine if a long word contains a
+     * zero byte. The method might have false positives but never false negatives.
+     * <p>
+     * This method is inspired by the `glibc/string/strlen.c` implementation
+     *
+     * @param segment    to examine
+     * @param fromOffset from where examination shall begin (inclusive)
+     * @param toOffset   to where examination shall end (exclusive)
+     * @throws IllegalArgumentException if the examined region contains no zero bytes
+     *                                  within a length that can be accepted by a String
+     */
+    @ForceInline
+    public static int strlenByte(final AbstractMemorySegmentImpl segment,
+                                 final long fromOffset,
+                                 final long toOffset) {
+        final long length = toOffset - fromOffset;
+        segment.checkBounds(fromOffset, length);
+        if (length == 0) {
+            // The state has to be checked explicitly for zero-length segments
+            segment.scope.checkValidState();
+            throw stringTooLarge(segment, fromOffset, toOffset);
+        }
+        final long longBytes = length & LONG_MASK;
+        final long longLimit = fromOffset + longBytes;
+        long offset = fromOffset;
+        for (; offset < longLimit; offset += Long.BYTES) {
+            long val = SCOPED_MEMORY_ACCESS.getLongUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset, !Architecture.isLittleEndian());
+            if (mightContainZeroByte(val)) {
+                for (int j = 0; j < Long.BYTES; j++) {
+                    if (SCOPED_MEMORY_ACCESS.getByte(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset + j) == 0) {
+                        return requireWithinArraySize(offset + j - fromOffset, segment, fromOffset, toOffset);
+                    }
+                }
+            }
+        }
+        // Handle the tail
+        for (; offset < toOffset; offset++) {
+            byte val = SCOPED_MEMORY_ACCESS.getByte(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset);
+            if (val == 0) {
+                return requireWithinArraySize(offset - fromOffset, segment, fromOffset, toOffset);
+            }
+        }
+        throw nullNotFound(segment, fromOffset, toOffset);
+    }
+
+    @ForceInline
+    public static int strlenShort(final AbstractMemorySegmentImpl segment,
+                                  final long fromOffset,
+                                  final long toOffset) {
+        final long length = toOffset - fromOffset;
+        segment.checkBounds(fromOffset, length);
+        if (length == 0) {
+            segment.scope.checkValidState();
+            throw stringTooLarge(segment, fromOffset, toOffset);
+        }
+        final long longBytes = length & LONG_MASK;
+        final long longLimit = fromOffset + longBytes;
+        long offset = fromOffset;
+        for (; offset < longLimit; offset += Long.BYTES) {
+            long val = SCOPED_MEMORY_ACCESS.getLongUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset, !Architecture.isLittleEndian());
+            if (mightContainZeroShort(val)) {
+                for (int j = 0; j < Long.BYTES; j += Short.BYTES) {
+                    if (SCOPED_MEMORY_ACCESS.getShortUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset + j, !Architecture.isLittleEndian()) == 0) {
+                        return requireWithinArraySize(offset + j - fromOffset, segment, fromOffset, toOffset);
+                    }
+                }
+            }
+        }
+        // Handle the tail
+        // Prevent over scanning as we step by 2
+        final long endScan = toOffset & ~1; // The last bit is zero
+        for (; offset < endScan; offset += Short.BYTES) {
+            short val = SCOPED_MEMORY_ACCESS.getShortUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset, !Architecture.isLittleEndian());
+            if (val == 0) {
+                return requireWithinArraySize(offset - fromOffset, segment, fromOffset, toOffset);
+            }
+        }
+        throw nullNotFound(segment, fromOffset, toOffset);
+    }
+
+    @ForceInline
+    public static int strlenInt(final AbstractMemorySegmentImpl segment,
+                                final long fromOffset,
+                                final long toOffset) {
+        final long length = toOffset - fromOffset;
+        segment.checkBounds(fromOffset, length);
+        if (length == 0) {
+            segment.scope.checkValidState();
+            throw stringTooLarge(segment, fromOffset, toOffset);
+        }
+        final long longBytes = length & LONG_MASK;
+        final long longLimit = fromOffset + longBytes;
+        long offset = fromOffset;
+        for (; offset < longLimit; offset += Long.BYTES) {
+            long val = SCOPED_MEMORY_ACCESS.getLongUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset, !Architecture.isLittleEndian());
+            if (mightContainZeroShort(val)) {
+                for (int j = 0; j < Long.BYTES; j += Integer.BYTES) {
+                    if (SCOPED_MEMORY_ACCESS.getIntUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset + j, !Architecture.isLittleEndian()) == 0) {
+                        return requireWithinArraySize(offset + j - fromOffset, segment, fromOffset, toOffset);
+                    }
+                }
+            }
+        }
+        // Handle the tail
+        // Prevent over scanning as we step by 4
+        final long endScan = toOffset & ~3; // The last two bit are zero
+        for (; offset < endScan; offset += Integer.BYTES) {
+            int val = SCOPED_MEMORY_ACCESS.getIntUnaligned(segment.sessionImpl(), segment.unsafeGetBase(), segment.unsafeGetOffset() + offset, !Architecture.isLittleEndian());
+            if (val == 0) {
+                return requireWithinArraySize(offset - fromOffset, segment, fromOffset, toOffset);
+            }
+        }
+        throw nullNotFound(segment, fromOffset, toOffset);
+    }
+
+    /*
+    Bits 63 and N * 8 (N = 1..7) of this number are zero.  Call these bits
+    the "holes".  Note that there is a hole just to the left of
+    each byte, with an extra at the end:
+
+    bits:  01111110 11111110 11111110 11111110 11111110 11111110 11111110 11111111
+    bytes: AAAAAAAA BBBBBBBB CCCCCCCC DDDDDDDD EEEEEEEE FFFFFFFF GGGGGGGG HHHHHHHH
+
+    The 1-bits make sure that carries propagate to the next 0-bit.
+    The 0-bits provide holes for carries to fall into.
+    */
+    private static final long HIMAGIC_FOR_BYTES = 0x8080_8080_8080_8080L;
+    private static final long LOMAGIC_FOR_BYTES = 0x0101_0101_0101_0101L;
+
+    private static boolean mightContainZeroByte(long l) {
+        return ((l - LOMAGIC_FOR_BYTES) & (~l) & HIMAGIC_FOR_BYTES) != 0;
+    }
+
+    private static final long HIMAGIC_FOR_SHORTS = 0x8000_8000_8000_8000L;
+    private static final long LOMAGIC_FOR_SHORTS = 0x0001_0001_0001_0001L;
+
+    static boolean mightContainZeroShort(long l) {
+        return ((l - LOMAGIC_FOR_SHORTS) & (~l) & HIMAGIC_FOR_SHORTS) != 0;
+    }
+
+
+    private static int requireWithinArraySize(long size,
+                                              AbstractMemorySegmentImpl segment,
+                                              long fromOffset,
+                                              long toOffset) {
+        if (size > ArraysSupport.SOFT_MAX_ARRAY_LENGTH) {
+            throw stringTooLarge(segment, fromOffset, toOffset);
+        }
+        return (int) size;
+    }
+
+    private static IllegalArgumentException stringTooLarge(AbstractMemorySegmentImpl segment,
+                                                           long fromOffset,
+                                                           long toOffset) {
+        return new IllegalArgumentException("String too large: " + exceptionInfo(segment, fromOffset, toOffset));
+    }
+
+    private static IndexOutOfBoundsException nullNotFound(AbstractMemorySegmentImpl segment,
+                                                          long fromOffset,
+                                                          long toOffset) {
+        return new IndexOutOfBoundsException("No null terminator found: " + exceptionInfo(segment, fromOffset, toOffset));
+    }
+
+    private static String exceptionInfo(AbstractMemorySegmentImpl segment,
+                                        long fromOffset,
+                                        long toOffset) {
+        return segment + " using region [" + fromOffset + ", " + toOffset + ")";
     }
 
     public enum CharsetKind {
